# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Brokerum API
  description: API для системы брокериджа заявок на кредитование и банковские гарантии
  version: 1.0.0
  contact:
    name: Brokerum Support
    email: support@brokerum.ru
    url: https://brokerum.ru
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.brokerum.ru/api
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /applications/{id}/files:
    get:
      summary: Получить файлы заявки
      description: Возвращает список файлов, прикрепленных к заявке
      tags:
        - Files
      parameters:
        - name: id
          in: path
          required: true
          description: ID заявки
          schema:
            type: integer
      responses:
        '200':
          description: Список файлов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /files/{id}:
    delete:
      summary: Удалить файл
      description: Удаляет файл
      tags:
        - Files
      parameters:
        - name: id
          in: path
          required: true
          description: ID файла
          schema:
            type: integer
      responses:
        '200':
          description: Файл удален
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      summary: Скачать файл
      description: Скачивает файл
      tags:
        - Files
      parameters:
        - name: id
          in: path
          required: true
          description: ID файла
          schema:
            type: integer
      responses:
        '200':
          description: Файл
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    File:
      type: object
      properties:
        id:
          type: integer
        application_id:
          type: integer
        file_name:
          type: string
        original_name:
          type: string
        file_path:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        file_hash:
          type: string
        uploaded_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean
        metadata:
          type: object
    FileListResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        total:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Описание ошибки"
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Bad Request"
  responses:
...
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...